refactor(evm): use native-width index and bound conversions
Per review: pcToEntry was indexed once with an nint (falling through to
the long index overload) while the other two reads narrowed to int, and
the (uint) bound guards truncated an nint rather than reinterpreting it.
Safe either way since the counter is bounded by the code length, but one
spelling is better in a PR about not paying for width conversions.

+++ src/Nethermind/Nethermind.Evm/VirtualMachine.DispatchSpecialized.std.cs
@@ -36,10 +36,10 @@ public unsafe partial class VirtualMachine<TGasPolicy>
         {
             int opCodeCount = 0;
             ref Instruction code = ref Unsafe.As<byte, Instruction>(ref stack.Code);
-            uint codeLength = (uint)stack.CodeLength;
+            nuint codeLength = (nuint)stack.CodeLength;
             // Hoisted: a no-op OnBeforeInstructionTrace would otherwise chase VmState.Env per instruction.
             int callDepth = VmState.Env.CallDepth;
-            while ((uint)pc < codeLength)
+            while ((nuint)pc < codeLength)
             {
 #if DEBUG
                 debugger?.TryWait(ref _currentState, ref pc, ref gas, ref stack.Head);
